resolve AF_INET/AF_INET6 platform dependency
AF_INET and AF_INET6 don't represent the same numeric values on
different platforms. Therefore we can't transmit them as part of
the target address from the client to the server. From now on we
use our own sibling constants instead.

This also prepares us for socks support in the future. The idea
would be, that the least significant bit indicates IPv4/v6. The
next bit indicates UDP or TCP. And the third bit could indicate
socks direction if required/implemented.